feat: add size in kb for the video meta.

+++ pkg/model/video.go
@@ -108,7 +108,7 @@ type UpdateVideoMeta struct {
 	IsFeatured      bool            `json:"is_featured,omitempty"`
 	Visibility      VideoVisibility `json:"visibility"`
 	Slug            string          `json:"slug"`
-	Size            uint64          `json:"size"`
+	Size            float32         `json:"size"`
 	Language        string          `json:"language"`
 	ResourceURL     url.URL         `json:"resource_url"`
 	StoragePath     string          `json:"-"`
